renaming to domains subdir of types
