+++ src/features/task/hooks/core/useTaskUpdate.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { TaskProps } from '@/shared/types/api.type';
 import { showSuccess } from '@/shared/services/notification/notificationService';
-import { UseUpdateTaskProps } from '@/shared/types/task/task.type';
+import { UseUpdateTaskProps } from '@/shared/types/domains/task.type';
 
 export const useTaskUpdate = ({
   initialData,
